@@ -28,68 +28,85 @@
     "**/*.{md,mdx,ts,js,tsx,jsx,json}": "npm run format"
   },
   "dependencies": {
-    "@apidevtools/json-schema-ref-parser": "^11.9.1",
-    "@docusaurus/core": "^3.7.0",
-    "@docusaurus/plugin-content-docs": "^3.7.0",
-    "@docusaurus/plugin-content-pages": "^3.7.0",
-    "@docusaurus/preset-classic": "^3.7.0",
-    "@docusaurus/theme-common": "^3.7.0",
-    "@docusaurus/theme-mermaid": "^3.7.0",
+    "@apidevtools/json-schema-ref-parser": "^14.2.0",
+    "@docusaurus/core": "3.9.2",
+    "@docusaurus/plugin-content-docs": "3.9.2",
+    "@docusaurus/plugin-content-pages": "3.9.2",
+    "@docusaurus/plugin-google-gtag": "3.9.2",
+    "@docusaurus/plugin-google-tag-manager": "3.9.2",
+    "@docusaurus/preset-classic": "3.9.2",
+    "@docusaurus/remark-plugin-npm2yarn": "3.9.2",
+    "@docusaurus/theme-common": "3.9.2",
+    "@docusaurus/theme-mermaid": "3.9.2",
     "@intercom/messenger-js-sdk": "^0.0.14",
+    "@lottiefiles/react-lottie-player": "^3.6.0",
     "@mdx-js/react": "^3.1.0",
-    "@metamask/design-tokens": "^5.0.0",
+    "@mermaid-js/layout-elk": "^0.1.9",
+    "@metamask/design-tokens": "^7.1.0",
     "@metamask/eth-sig-util": "^7.0.3",
-    "@metamask/profile-sync-controller": "^4.1.1",
-    "@metamask/sdk": "^0.32.0",
-    "@rjsf/core": "^5.24.1",
-    "@rjsf/utils": "^5.24.3",
-    "@rjsf/validator-ajv8": "^5.24.2",
-    "@sentry/browser": "^8.51.0",
+    "@metamask/profile-sync-controller": "^16.0.0",
+    "@metamask/sdk": "^0.33.1",
+    "@rjsf/core": "^5.24.12",
+    "@rjsf/utils": "^5.24.13",
+    "@rjsf/validator-ajv8": "^5.24.12",
+    "@sentry/browser": "^10.22.0",
     "@types/react": "^18.3.3",
+    "classnames": "^2.5.1",
     "clsx": "^2.1.1",
+    "copy-to-clipboard": "^3.3.3",
     "docusaurus-plugin-sass": "^0.2.5",
-    "docusaurus-plugin-segment": "^1.0.4",
-    "dotenv": "^16.4.5",
-    "ethers": "^6.13.5",
+    "dotenv": "^17.2.1",
+    "ethers": "^6.15.0",
     "js-cookie": "^3.0.5",
     "jsonwebtoken": "^9.0.2",
-    "launchdarkly-js-client-sdk": "^3.3.0",
+    "katex": "^0.16.25",
+    "launchdarkly-js-client-sdk": "^3.9.0",
     "lodash.camelcase": "^4.3.0",
     "lodash.debounce": "^4.0.8",
     "lodash.isplainobject": "^4.0.6",
     "node-polyfill-webpack-plugin": "^2.0.1",
-    "prettier": "^3.3.3",
-    "prism-react-renderer": "^2.1.0",
+    "prettier": "^3.6.2",
+    "prism-react-renderer": "^2.4.1",
     "react": "^18.0.0",
     "react-alert": "^7.0.3",
+    "react-bookmark": "^0.8.2",
     "react-dom": "^18.0.0",
-    "react-dropdown-select": "^4.11.4",
+    "react-dropdown-select": "^4.12.2",
+    "react-icons": "^5.5.0",
     "react-modal": "^3.16.3",
-    "react-player": "^2.13.0",
+    "react-player": "^3.3.1",
+    "react-select": "^5.10.2",
+    "react-spinners": "^0.17.0",
+    "react-spring": "^10.0.3",
     "react-tippy": "^1.4.0",
+    "rehype-katex": "^7.0.1",
     "remark-codesandbox": "^0.10.1",
     "remark-docusaurus-tabs": "^0.2.0",
-    "sass": "^1.83.4"
+    "remark-math": "^6.0.0",
+    "sass": "^1.93.0",
+    "docusaurus-plugin-llms": "^0.2.2"
   },
   "devDependencies": {
-    "@docusaurus/eslint-plugin": "^3.7.0",
-    "@docusaurus/module-type-aliases": "^3.7.0",
-    "@docusaurus/plugin-client-redirects": "^3.7.0",
-    "@docusaurus/tsconfig": "^3.7.0",
-    "@docusaurus/types": "^3.7.0",
-    "@eslint/js": "^9.21.0",
-    "@tsconfig/docusaurus": "^2.0.3",
+    "@docusaurus/eslint-plugin": "3.9.2",
+    "@docusaurus/module-type-aliases": "3.9.2",
+    "@docusaurus/plugin-client-redirects": "3.9.2",
+    "@docusaurus/tsconfig": "3.9.2",
+    "@docusaurus/types": "3.9.2",
+    "@eslint/js": "^9.36.0",
+    "@tsconfig/docusaurus": "^2.0.7",
     "@types/react-alert": "^7.0.6",
     "@types/react-modal": "^3.16.3",
-    "@typescript-eslint/parser": "^8.26.1",
-    "eslint": "^9.20.0",
-    "eslint-plugin-react": "^7.37.4",
+    "@typescript-eslint/parser": "^8.46.2",
+    "eslint": "^9.35.0",
+    "eslint-plugin-react": "^7.37.5",
     "stylelint": "^15.0.0 ",
     "stylelint-config-standard": "^34.0.0",
     "tsc-files": "^1.1.4",
-    "typescript": "^5.5.4"
+    "typescript": "^5.8.3"
   },
   "overrides": {
+    "axios": "^1.7.9",
+    "cross-spawn": "^7.0.6",
     "react-alert": {
       "react": "^18.0.0",
       "react-dom": "^18.0.0"
@@ -108,6 +125,6 @@
     ]
   },
   "engines": {
-    "node": ">=18"
+    "node": "20.x"
   }
 }
