@@ -19,6 +19,7 @@ export default [{
         "**/.eslintrc.js",
         "node_modules/*",
         ".docusaurus/*",
+        ".integrationBuilderCache/*",
         "**/*.md",
         "**/*.mdx",
         "static/*",
@@ -41,7 +42,13 @@ export default [{
             },
 
             project: ["./tsconfig.json"],
-            tsconfigRootDir: "./",
+            tsconfigRootDir: __dirname,
+        },
+    },
+
+    settings: {
+        react: {
+            version: "detect",
         },
     },
 
