@@ -1,50 +1,68 @@
 import clsx from 'clsx'
-import { CSSProperties, JSX, useState, useEffect } from 'react'
+import { CSSProperties, JSX } from 'react'
 import { useColorMode } from '@docusaurus/theme-common'
-import Heading from '@theme/Heading'
 import Card, { CardItem } from '@site/src/components/Card'
 
 import styles from './CardSection.module.scss'
 
 type CardSectionProps = {
+  title?: string
+  description?: string
   cards: CardItem[]
-  colorPalette: string
+  colorPalette?: string
 }
 
-export default function CardSection({ cards, colorPalette }: CardSectionProps): JSX.Element {
+export default function CardSection({
+  title,
+  description,
+  cards,
+  colorPalette,
+}: CardSectionProps): JSX.Element {
   const { colorMode } = useColorMode()
-  const [theme, setTheme] = useState('')
-
-  useEffect(() => {
-    setTheme(colorMode)
-  }, [colorMode])
+  const theme = colorMode
 
   return (
-    <section className={styles['wrapper']}>
+    <section className={styles.wrapper}>
       <div className="container">
         <div className={styles['grid-wrapper']}>
           <div className={styles['grid-col-center']}>
-            <ul
-              className={styles['grid']}
+            <div
+              className={clsx(
+                styles['section-grid'],
+                !title && !description && styles['cards-only']
+              )}
               style={
                 colorPalette
                   ? ({
                       '--color-palette': `var(--developer-${colorPalette})`,
                     } as CSSProperties)
                   : {}
-              }
-            >
-              {cards?.length > 0 &&
-                cards.map(({ title, description, href }: CardItem, cardIndex: number) => (
-                  <Card
-                    key={cardIndex}
-                    title={title}
-                    description={description}
-                    href={href}
-                    theme={theme}
-                  />
+              }>
+              {/* Title and Description Column */}
+              {(title || description) && (
+                <div className={styles['content-column']}>
+                  <div className={styles.header}>
+                    {title && <h2>{title}</h2>}
+                    {description && <p>{description}</p>}
+                  </div>
+                </div>
+              )}
+
+              {/* Cards Columns */}
+              <div className={styles['cards-wrapper']}>
+                {cards.map(({ title, description, href, buttonIcon }, index) => (
+                  <div key={index} className={styles['card-column']}>
+                    <Card
+                      title={title}
+                      description={description}
+                      href={href}
+                      theme={theme}
+                      buttonIcon={buttonIcon}
+                    />
+                  </div>
                 ))}
-            </ul>
+              </div>
+            </div>
           </div>
         </div>
       </div>
