@@ -3,7 +3,6 @@ import Text from '@site/src/components/Text'
 import Button from '@site/src/components/elements/buttons/button'
 import Input from '@site/src/components/Input'
 import clsx from 'clsx'
-import { trackClickForSegment } from '@site/src/lib/segmentAnalytics'
 import { WALLET_LINK_TYPE } from '@site/src/components/AuthLogin/AuthModal'
 import { MetamaskProviderContext } from '@site/src/theme/Root'
 import styles from './hero.module.scss'
@@ -18,6 +17,7 @@ interface IHero {
   inputValue?: string
   isLoading?: boolean
   isLimitedUserPlan?: boolean
+  isMaintenance?: boolean
 }
 
 export default function Hero({
@@ -28,6 +28,7 @@ export default function Hero({
   handleOnInputChange,
   isLoading,
   isLimitedUserPlan,
+  isMaintenance,
 }: IHero) {
   const {
     metaMaskAccount,
@@ -46,42 +47,15 @@ export default function Hero({
 
   const handleConnectWallet = () => {
     setIsWalletLinking(true)
-    trackClickForSegment({
-      eventName: !showInstallButton ? 'Install MetaMask' : 'Connect MetaMask',
-      clickType: 'Hero',
-      userExperience: 'B',
-      responseStatus: null,
-      responseMsg: null,
-      timestamp: Date.now(),
-    })
     metaMaskWalletIdConnectHandler()
   }
 
   const handleLinkWallet = () => {
     setIsWalletLinking(true)
-    trackClickForSegment({
-      eventName:
-        walletLinked === WALLET_LINK_TYPE.NO
-          ? 'Link Developer Dashboard Account'
-          : 'Select Developer Dashboard Account',
-      clickType: 'Hero',
-      userExperience: 'B',
-      responseStatus: null,
-      responseMsg: null,
-      timestamp: Date.now(),
-    })
     window.location.href = walletAuthUrl
   }
 
   const handleRequestEth = () => {
-    trackClickForSegment({
-      eventName: 'Request ETH',
-      clickType: 'Hero',
-      userExperience: 'B',
-      responseStatus: null,
-      responseMsg: null,
-      timestamp: Date.now(),
-    })
     handleRequest()
   }
 
@@ -118,10 +92,12 @@ export default function Hero({
                     : walletLinked === WALLET_LINK_TYPE.NO
                       ? 'Link your Developer Dashboard account to get started and request ETH.'
                       : 'Select your Developer Dashboard account to get started and request ETH.'
-                : 'Enter your MetaMask wallet address and request ETH.'}
+                : !isMaintenance
+                  ? 'Enter your MetaMask wallet address and request ETH.'
+                  : 'The faucet is at full capacity due to high demand. Try checking back later.'}
           </Text>
           <div className={styles.actions}>
-            {!!Object.keys(projects).length && !showInstallButton && (
+            {!!Object.keys(projects).length && !showInstallButton && !isMaintenance && (
               <div className={styles.inputCont}>
                 <div className={styles.inputWrapper}>
                   <Input
