@@ -1,10 +1,10 @@
-export { default as Faq } from "./Faq";
-export { default as TransactionTable } from "./TransactionTable";
-export { default as Hero } from "./Hero";
-export { default as Maintenance } from "./Maintenance";
+export { default as Faq } from './Faq'
+export { default as TransactionTable } from './TransactionTable'
+export { default as Hero } from './Hero'
 export {
   AlertCommonIssue,
   AlertPastActivity,
   AlertCooldown,
   AlertSuccess,
-} from "./Alerts";
+  AlertFailedPoh,
+} from './Alerts'
