@@ -1,24 +1,28 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react'
+
+interface NetworkData {
+  [chainId: string]: string
+}
 
 const GasApiNetworks = () => {
-  const [networks, setNetworks] = useState({});
+  const [networks, setNetworks] = useState<NetworkData>({})
 
   useEffect(() => {
     const fetchData = async () => {
       try {
-        const response = await fetch('https://gas.api.cx.metamask.io/v1/supportedChainNames');
-        const data = await response.json();
-        setNetworks(data);
+        const response = await fetch('https://gas.api.cx.metamask.io/v1/supportedChainNames')
+        const data: NetworkData = await response.json()
+        setNetworks(data)
       } catch (error) {
-        console.error('Error fetching data:', error);
+        console.error('Error fetching data:', error)
       }
-    };
+    }
 
-    fetchData();
-  }, []);
+    fetchData()
+  }, [])
 
   // Convert the networks object into an array of [chainId, networkName] pairs
-  const networksArray = Object.entries(networks);
+  const networksArray = Object.entries(networks)
 
   return (
     <table>
@@ -37,7 +41,7 @@ const GasApiNetworks = () => {
         ))}
       </tbody>
     </table>
-  );
-};
+  )
+}
 
-export default GasApiNetworks;
+export default GasApiNetworks
