@@ -6,7 +6,7 @@ import Shape from '@site/static/img/shapes/hero/shape.svg'
 
 import styles from './Hero.module.scss'
 
-export default function Hero({ title, description, button = null }) {
+export default function Hero({ title, description, buttons = [] }) {
   return (
     <section className={styles['wrapper']}>
       <div className="container">
@@ -20,17 +20,26 @@ export default function Hero({ title, description, button = null }) {
               <p className={clsx(styles['description'], 'type-paragraph-l')}>{description}</p>
             )}
 
-            {button && (
-              <Button
-                className={styles['button']}
-                as="link"
-                type="secondary"
-                {...button}
-                style={{
-                  '--button-color-hover': 'var(--general-white)',
-                  '--button-text-color-hover': 'var(--general-black)',
-                }}
-              />
+            {buttons.length > 0 && (
+              <div className={styles['buttons-container']}>
+                {buttons.map((button, index) => (
+                  <Button
+                    key={index}
+                    className={styles['button']}
+                    as="link"
+                    type="primary"
+                    {...button}
+                    style={{
+                      '--button-background': 'var(--general-white)',
+                      '--button-color': 'var(--general-black)',
+                      '--button-border-color': 'var(--general-white)',
+                      '--button-background-hover': 'var(--general-black)',
+                      '--button-color-hover': 'var(--general-white)',
+                      '--button-border-color-hover': 'var(--general-white)',
+                    }}
+                  />
+                ))}
+              </div>
             )}
           </div>
           <div className={styles['grid-col-right']}>
