@@ -1,16 +1,16 @@
-import React from "react";
-import clsx from "clsx";
+import React from 'react'
+import clsx from 'clsx'
 
-import styles from "./input.module.scss";
+import styles from './input.module.scss'
 
 interface IInput {
-  onChange?: (string) => void;
-  disabled?: boolean;
-  label?: string;
-  className?: string;
-  error?: string;
-  placeholder?: string;
-  value?: string;
+  onChange?: (string) => void
+  disabled?: boolean
+  label?: string
+  className?: string
+  error?: string
+  placeholder?: string
+  value?: string
 }
 
 export default function Input({
@@ -31,9 +31,9 @@ export default function Input({
         disabled={disabled}
         placeholder={placeholder}
         className={clsx(styles.input, error && styles.error)}
-        onChange={(e) => onChange(e?.target?.value)}
+        onChange={e => onChange(e?.target?.value)}
       />
       {error && <p className={styles.errorMsg}>{error}</p>}
     </label>
-  );
+  )
 }
