@@ -12,6 +12,7 @@ interface SchemaPropertyProps {
   description?: string
   pattern?: string
   defaultVal?: string
+  showRequired?: boolean
 }
 
 interface TagProps {
@@ -25,6 +26,7 @@ export const SchemaProperty = ({
   description,
   pattern,
   defaultVal,
+  showRequired = true,
 }: SchemaPropertyProps) => {
   return (
     <div className={styles.schemaWrapper}>
@@ -34,19 +36,19 @@ export const SchemaProperty = ({
             {title}
           </Heading>
           <span className={clsx(styles.textAltColor, 'type-paragraph-m')}>{type}</span>
+          {pattern && (
+            <div className={styles.patternWrapper}>
+              <span className={styles.patternLabel}>Pattern: </span>
+              <span className={styles.patternValue}>{pattern}</span>
+            </div>
+          )}
         </div>
-        {required && (
+        {required && showRequired && (
           <span className={clsx(styles.textErrorColor, 'type-paragraph-m')}>required</span>
         )}
       </div>
       <p className={clsx(styles.description, 'margin--none type-paragraph-m')}>
         <MDContent content={description} />
-        {pattern && (
-          <div className={styles.propItemWrapper}>
-            <span className={styles.propItemLabel}>Pattern: </span>
-            <span className={styles.propItemValue}>{pattern}</span>
-          </div>
-        )}
         {defaultVal && (
           <div className={styles.propItemWrapper}>
             <span className={styles.propItemLabel}>Default: </span>
