@@ -1,49 +1,45 @@
-import React, { useContext, useEffect, useState } from "react";
-import { FieldTemplateProps } from "@rjsf/utils";
-import { BaseInputTemplate } from "@site/src/components/ParserOpenRPC/InteractiveBox/templates/BaseInputTemplate";
-import { SelectWidget } from "@site/src/components/ParserOpenRPC/InteractiveBox/widgets/SelectWidget";
-import styles from "@site/src/components/ParserOpenRPC/InteractiveBox/styles.module.scss";
-import { ParserOpenRPCContext } from "@site/src/components/ParserOpenRPC";
-import clsx from "clsx";
+import React, { useContext, useEffect, useState } from 'react'
+import { FieldTemplateProps } from '@rjsf/utils'
+import { BaseInputTemplate } from '@site/src/components/ParserOpenRPC/InteractiveBox/templates/BaseInputTemplate'
+import { SelectWidget } from '@site/src/components/ParserOpenRPC/InteractiveBox/widgets/SelectWidget'
+import styles from '@site/src/components/ParserOpenRPC/InteractiveBox/styles.module.scss'
+import { ParserOpenRPCContext } from '@site/src/components/ParserOpenRPC'
+import clsx from 'clsx'
 
 export const ConditionalField = (props: FieldTemplateProps) => {
-  const [isOpened, setIsOpened] = useState(false);
-  const [selectedTypeSchema, setSelectedTypeSchema] = useState(null);
-  const [isEditView, setIsEditView] = useState(false);
-  const {
-    setIsDrawerContentFixed,
-    setDrawerLabel,
-    isComplexTypeView,
-    setIsComplexTypeView,
-  } = useContext(ParserOpenRPCContext);
+  const [isOpened, setIsOpened] = useState(false)
+  const [selectedTypeSchema, setSelectedTypeSchema] = useState(null)
+  const [isEditView, setIsEditView] = useState(false)
+  const { setIsDrawerContentFixed, setDrawerLabel, isComplexTypeView, setIsComplexTypeView } =
+    useContext(ParserOpenRPCContext)
   // @ts-ignore
-  const { formData, schema, name, onChange } = props;
-  const listItems = schema?.anyOf ? schema?.anyOf : schema?.oneOf;
-  const checkForNullTypeSchema = (type) => type === "null";
+  const { formData, schema, name, onChange } = props
+  const listItems = schema?.anyOf ? schema?.anyOf : schema?.oneOf
+  const checkForNullTypeSchema = type => type === 'null'
   const showComplexTypeView = () => {
-    setDrawerLabel(name);
-    setIsDrawerContentFixed(true);
-    setIsEditView(true);
-    setIsComplexTypeView(true);
-  };
-  const onDropdownOptionClick = (e) => {
+    setDrawerLabel(name)
+    setIsDrawerContentFixed(true)
+    setIsEditView(true)
+    setIsComplexTypeView(true)
+  }
+  const onDropdownOptionClick = e => {
     const selectedSchema = listItems.find(
       // @ts-ignore
-      ({ title }) => title === e.target.dataset.value,
-    );
+      ({ title }) => title === e.target.dataset.value
+    )
     // @ts-ignore
-    const isNullTypeSchema = checkForNullTypeSchema(selectedSchema?.type);
+    const isNullTypeSchema = checkForNullTypeSchema(selectedSchema?.type)
     if (isNullTypeSchema) {
-      onChange(null);
+      onChange(null)
     } else {
       setSelectedTypeSchema(
         // @ts-ignore
-        listItems.find(({ title }) => title === e.target.dataset.value),
-      );
-      showComplexTypeView();
+        listItems.find(({ title }) => title === e.target.dataset.value)
+      )
+      showComplexTypeView()
     }
-    setIsOpened(false);
-  };
+    setIsOpened(false)
+  }
   const selectWidgetProps = {
     ...props,
     schema: selectedTypeSchema,
@@ -51,24 +47,24 @@ export const ConditionalField = (props: FieldTemplateProps) => {
     value: formData,
     ...(selectedTypeSchema?.enum && {
       options: {
-        enumOptions: selectedTypeSchema?.enum.map((item) => ({
+        enumOptions: selectedTypeSchema?.enum.map(item => ({
           label: item,
           value: item,
         })),
       },
     }),
-  };
+  }
   const baseInputProps = {
     ...props,
     schema: selectedTypeSchema,
-  };
+  }
 
   useEffect(() => {
     if (!isComplexTypeView) {
-      setIsEditView(false);
-      setSelectedTypeSchema(null);
+      setIsEditView(false)
+      setSelectedTypeSchema(null)
     }
-  }, [isComplexTypeView]);
+  }, [isComplexTypeView])
 
   return listItems?.length > 0 ? (
     <>
@@ -77,51 +73,41 @@ export const ConditionalField = (props: FieldTemplateProps) => {
           <label className={styles.tableColumnParam}>{name}</label>
         </div>
         <div className={styles.tableColumn}>
-          <div
-            className={clsx(styles.tableValueRow, styles.tableValueRowPadding)}
-          >
-            <div className={styles.tableValueConditionalField}>{formData === undefined ? "" : String(formData)}</div>
+          <div className={clsx(styles.tableValueRow, styles.tableValueRowPadding)}>
+            <div className={styles.tableValueConditionalField}>
+              {formData === undefined ? '' : String(formData)}
+            </div>
             <span
-              className={clsx(
-                styles.tableColumnType,
-                styles.tableColumnTypeDropdown,
-              )}
+              className={clsx(styles.tableColumnType, styles.tableColumnTypeDropdown)}
               onClick={() => {
-                setIsOpened(!isOpened);
-              }}
-            >
+                setIsOpened(!isOpened)
+              }}>
               <span className={styles.dropdown}>
-                {schema?.anyOf ? "anyOf" : "oneOf"}
+                {schema?.anyOf ? 'anyOf' : 'oneOf'}
                 <span
                   className={clsx(
                     styles.tableColumnIcon,
                     styles.chevronIcon,
                     styles.dropdownChevronIcon,
-                    !isOpened && styles.chevronIconDown,
+                    !isOpened && styles.chevronIconDown
                   )}
                 />
                 <span
                   className={clsx(
                     styles.chevronIcon,
                     styles.dropdownChevronIcon,
-                    !isOpened && styles.chevronIconDown,
+                    !isOpened && styles.chevronIconDown
                   )}
                 />
               </span>
-              <ul
-                className={clsx(
-                  styles.dropdownList,
-                  !isOpened && styles.dropdownListClosed,
-                )}
-              >
+              <ul className={clsx(styles.dropdownList, !isOpened && styles.dropdownListClosed)}>
                 {listItems?.map((listItem, index) => (
                   <li
                     className={styles.dropdownItem}
                     key={index}
                     onClick={onDropdownOptionClick}
                     // @ts-ignore
-                    data-value={listItem.title}
-                  >
+                    data-value={listItem.title}>
                     {/* @ts-ignore */}
                     {listItem.title}
                   </li>
@@ -134,7 +120,7 @@ export const ConditionalField = (props: FieldTemplateProps) => {
       {isComplexTypeView &&
       isEditView &&
       selectedTypeSchema &&
-      selectedTypeSchema.type !== "null" ? (
+      selectedTypeSchema.type !== 'null' ? (
         <div className={styles.tableComplexType}>
           {selectedTypeSchema?.enum ? (
             // @ts-ignore
@@ -146,5 +132,5 @@ export const ConditionalField = (props: FieldTemplateProps) => {
         </div>
       ) : null}
     </>
-  ) : null;
-};
+  ) : null
+}
