@@ -1,16 +1,10 @@
-import React, { useState } from "react";
-import { WidgetProps } from "@rjsf/utils";
-import clsx from "clsx";
-import styles from "@site/src/components/ParserOpenRPC/InteractiveBox/styles.module.scss";
+import React, { useState } from 'react'
+import { WidgetProps } from '@rjsf/utils'
+import clsx from 'clsx'
+import styles from '@site/src/components/ParserOpenRPC/InteractiveBox/styles.module.scss'
 
-export const DropdownWidget = ({
-  name,
-  value,
-  onChange,
-  schema,
-  options,
-}: WidgetProps) => {
-  const [isOpened, setIsOpened] = useState(false);
+export const DropdownWidget = ({ name, value, onChange, schema, options }: WidgetProps) => {
+  const [isOpened, setIsOpened] = useState(false)
 
   return (
     <div className={styles.tableRow}>
@@ -18,19 +12,13 @@ export const DropdownWidget = ({
         <label className={styles.tableColumnParam}>{name}</label>
       </div>
       <div className={styles.tableColumn}>
-        <div
-          className={clsx(styles.tableValueRow, styles.tableValueRowPadding)}
-        >
-          {value === undefined ? "" : String(value)}
+        <div className={clsx(styles.tableValueRow, styles.tableValueRowPadding)}>
+          {value === undefined ? '' : String(value)}
           <span
-            className={clsx(
-              styles.tableColumnType,
-              styles.tableColumnTypeDropdown,
-            )}
+            className={clsx(styles.tableColumnType, styles.tableColumnTypeDropdown)}
             onClick={() => {
-              setIsOpened(!isOpened);
-            }}
-          >
+              setIsOpened(!isOpened)
+            }}>
             <span className={styles.dropdown}>
               {schema.type}
               <span
@@ -38,32 +26,26 @@ export const DropdownWidget = ({
                   styles.tableColumnIcon,
                   styles.chevronIcon,
                   styles.dropdownChevronIcon,
-                  !isOpened && styles.chevronIconDown,
+                  !isOpened && styles.chevronIconDown
                 )}
               />
               <span
                 className={clsx(
                   styles.chevronIcon,
                   styles.dropdownChevronIcon,
-                  !isOpened && styles.chevronIconDown,
+                  !isOpened && styles.chevronIconDown
                 )}
               />
             </span>
-            <ul
-              className={clsx(
-                styles.dropdownList,
-                !isOpened && styles.dropdownListClosed,
-              )}
-            >
+            <ul className={clsx(styles.dropdownList, !isOpened && styles.dropdownListClosed)}>
               {options.enumOptions.map(({ value }, index) => (
                 <li
                   className={styles.dropdownItem}
                   key={index}
                   onClick={() => {
-                    onChange(value);
-                    setIsOpened(false);
-                  }}
-                >
+                    onChange(value)
+                    setIsOpened(false)
+                  }}>
                   {String(value)}
                 </li>
               ))}
@@ -72,5 +54,5 @@ export const DropdownWidget = ({
         </div>
       </div>
     </div>
-  );
-};
+  )
+}
