@@ -1,12 +1,12 @@
-import React, { useContext, useEffect, useState } from "react";
-import ldClient from "launchdarkly";
-import { MetamaskProviderContext } from "@site/src/theme/Root";
-import Select from "react-dropdown-select";
-import Button from "@site/src/components/Button";
-import styles from "./styles.module.scss";
-import { WALLET_LINK_TYPE } from "@site/src/components/AuthLogin/AuthModal";
+import React, { useContext, useEffect, useState } from 'react'
+import ldClient from 'launchdarkly'
+import { MetamaskProviderContext } from '@site/src/theme/Root'
+import Select from 'react-dropdown-select'
+import Button from '@site/src/components/Button'
+import styles from './styles.module.scss'
+import { WALLET_LINK_TYPE } from '@site/src/components/AuthLogin/AuthModal'
 
-const LOGIN_FF = "mm-unified-login";
+const LOGIN_FF = 'mm-unified-login'
 
 const ProjectsBox = () => {
   const {
@@ -17,53 +17,51 @@ const ProjectsBox = () => {
     metaMaskWalletIdConnectHandler,
     walletAuthUrl,
     setUserAPIKey,
-  } = useContext(MetamaskProviderContext);
-  const options = Object.keys(projects).map((v) => ({
+  } = useContext(MetamaskProviderContext)
+  const options = Object.keys(projects).map(v => ({
     value: v,
     label: projects[v].name,
-  }));
-  const [currentProject, setCurrentProject] = useState(
-    [options[0]].filter(Boolean)
-  );
-  const [ldReady, setLdReady] = useState(false);
-  const [loginEnabled, setLoginEnabled] = useState(false);
-  const [isWalletLinking, setIsWalletLinking] = useState(false);
+  }))
+  const [currentProject, setCurrentProject] = useState([options[0]].filter(Boolean))
+  const [ldReady, setLdReady] = useState(false)
+  const [loginEnabled, setLoginEnabled] = useState(false)
+  const [isWalletLinking, setIsWalletLinking] = useState(false)
 
   const metaMaskWalletConnectionHandler = () => {
-    setIsWalletLinking(true);
-    metaMaskWalletIdConnectHandler();
+    setIsWalletLinking(true)
+    metaMaskWalletIdConnectHandler()
   }
 
   useEffect(() => {
     ldClient.waitUntilReady().then(() => {
-      setLoginEnabled(ldClient.variation(LOGIN_FF, false));
-      setLdReady(true);
-    });
-    const handleChange = (current) => {
-      setLoginEnabled(current);
-    };
-    ldClient.on(`change:${LOGIN_FF}`, handleChange);
+      setLoginEnabled(ldClient.variation(LOGIN_FF, false))
+      setLdReady(true)
+    })
+    const handleChange = current => {
+      setLoginEnabled(current)
+    }
+    ldClient.on(`change:${LOGIN_FF}`, handleChange)
     return () => {
-      ldClient.off(`change:${LOGIN_FF}`, handleChange);
-    };
-  }, []);
+      ldClient.off(`change:${LOGIN_FF}`, handleChange)
+    }
+  }, [])
 
   useEffect(() => {
     if (!currentProject.length && options[0]) {
-      setCurrentProject([options[0]]);
-      setUserAPIKey(options[0].value);
+      setCurrentProject([options[0]])
+      setUserAPIKey(options[0].value)
     }
-  }, [projects]);
+  }, [projects])
 
   useEffect(() => {
     if (options?.length > 0) {
-      setUserAPIKey(options[0].value);
+      setUserAPIKey(options[0].value)
     }
-  }, [options.length]);
+  }, [options.length])
 
   useEffect(() => {
     if (walletLinked) {
-      setIsWalletLinking(false);
+      setIsWalletLinking(false)
     }
   }, [walletLinked])
 
@@ -80,37 +78,36 @@ const ProjectsBox = () => {
             searchable={false}
             options={options}
             values={currentProject}
-            onChange={(value) => {
-              setCurrentProject(value);
-              setUserAPIKey(value[0].value);
+            onChange={value => {
+              setCurrentProject(value)
+              setUserAPIKey(value[0].value)
             }}
             contentRenderer={({ state }) => {
               return (
                 <div>
-                  {state.values.map((item) => (
+                  {state.values.map(item => (
                     <div key={item.value}>
                       <div className={styles.selectDropdownLabel}>{item.label}</div>
                       <div className={styles.selectDropdownValue}>{item.value}</div>
                     </div>
                   ))}
                 </div>
-              );
+              )
             }}
             dropdownRenderer={({ methods }) => {
               return (
                 <div className={styles.selectDropdown}>
-                  {options.map((option) => (
+                  {options.map(option => (
                     <div
                       className={styles.selectDropdownOption}
                       key={option.value}
-                      onClick={() => methods.addItem(option)}
-                    >
+                      onClick={() => methods.addItem(option)}>
                       <div className={styles.selectDropdownLabel}>{option.label}</div>
                       <div className={styles.selectDropdownValue}>{option.value}</div>
                     </div>
                   ))}
                 </div>
-              );
+              )
             }}
           />
         ) : (
@@ -118,62 +115,50 @@ const ProjectsBox = () => {
             {!needsMfa && walletLinked === undefined && (
               <>
                 <div>
-                  {isWalletLinking ?
-                    "Don’t close or exit this page. Please continue connecting on your extension." :
-                    "Connect your MetaMask wallet to start sending requests to your Infura API keys."
-                  }
+                  {isWalletLinking
+                    ? 'Don’t close or exit this page. Please continue connecting on your extension.'
+                    : 'Connect your MetaMask wallet to start sending requests to your Infura API keys.'}
                 </div>
                 <Button
                   thin
                   className={styles.connectButton}
                   onClick={metaMaskWalletConnectionHandler}
-                  isLoading={isWalletLinking}
-                >
+                  isLoading={isWalletLinking}>
                   Connect MetaMask
                 </Button>
               </>
             )}
             {needsMfa && (
               <>
-                <div>
-                  Your Infura account needs two factor auth token
-                </div>
+                <div>Your Infura account needs two factor auth token</div>
                 <Button
                   thin
                   variant="secondary"
                   wrapText={false}
-                  onClick={() => (window.location.href = walletAuthUrl)}
-                >
+                  onClick={() => (window.location.href = walletAuthUrl)}>
                   Enter your MFA
                 </Button>
               </>
             )}
             {walletLinked === WALLET_LINK_TYPE.NO && (
               <>
-                <div>
-                  Link your Infura account to send requests to your Infura API
-                  keys.
-                </div>
+                <div>Link your Infura account to send requests to your Infura API keys.</div>
                 <Button
                   thin
                   variant="secondary"
                   wrapText={false}
-                  onClick={() => (window.location.href = walletAuthUrl)}
-                >
+                  onClick={() => (window.location.href = walletAuthUrl)}>
                   Link Infura Account
                 </Button>
               </>
             )}
             {walletLinked === WALLET_LINK_TYPE.MULTIPLE && (
               <>
-                <div>
-                  Select MetaMask Developer account linked with your current wallet.
-                </div>
+                <div>Select MetaMask Developer account linked with your current wallet.</div>
                 <Button
                   thin
                   className={styles.connectButton}
-                  onClick={() => (window.location.href = walletAuthUrl)}
-                >
+                  onClick={() => (window.location.href = walletAuthUrl)}>
                   Select MetaMask Developer Account
                 </Button>
               </>
@@ -182,7 +167,7 @@ const ProjectsBox = () => {
         )}
       </div>
     )
-  );
-};
+  )
+}
 
-export default ProjectsBox;
+export default ProjectsBox
