@@ -1,15 +1,15 @@
-import React, { ReactNode } from "react";
-import styles from "./tooltip.module.css";
-import { Tooltip as ReactTippy } from "react-tippy";
-import "react-tippy/dist/tippy.css";
-import clsx from "clsx";
+import React, { ReactNode } from 'react'
+import styles from './tooltip.module.css'
+import { Tooltip as ReactTippy } from 'react-tippy'
+import 'react-tippy/dist/tippy.css'
+import clsx from 'clsx'
 
 interface TooltipProps {
-  children: ReactNode;
-  message: string;
-  disabled?: boolean;
-  className?: string;
-  onHidden?: VoidFunction;
+  children: ReactNode
+  message: string
+  disabled?: boolean
+  className?: string
+  onHidden?: VoidFunction
 }
 
 export const Tooltip = ({
@@ -25,15 +25,10 @@ export const Tooltip = ({
       disabled={disabled}
       arrow={true}
       onHidden={onHidden}
-      html={
-        <div className={clsx(styles.tooltipContainer, className)}>
-          {message}
-        </div>
-      }
+      html={<div className={clsx(styles.tooltipContainer, className)}>{message}</div>}
       position="top"
-      interactive={true}
-    >
+      interactive={true}>
       {children}
     </ReactTippy>
   </>
-);
+)
