@@ -1,61 +1,56 @@
 declare global {
   interface WindowEventMap {
     // @ts-ignore
-    "eip6963:announceProvider": CustomEvent;
+    'eip6963:announceProvider': CustomEvent
   }
 }
 
 export interface EIP6963ProviderInfo {
-  walletId: string;
-  uuid: string;
-  name: string;
-  icon: string;
+  walletId: string
+  uuid: string
+  name: string
+  icon: string
 }
 
 export interface EIP1193Provider {
-  isStatus?: boolean;
-  host?: string;
-  path?: string;
+  isStatus?: boolean
+  host?: string
+  path?: string
   sendAsync?: (
     request: { method: string; params?: Array<unknown> },
-    callback: (error: Error | null, response: unknown) => void,
-  ) => void;
+    callback: (error: Error | null, response: unknown) => void
+  ) => void
   send?: (
     request: { method: string; params?: Array<unknown> },
-    callback: (error: Error | null, response: unknown) => void,
-  ) => void;
-  request: (request: {
-    method: string;
-    params?: Array<unknown>;
-  }) => Promise<unknown>;
+    callback: (error: Error | null, response: unknown) => void
+  ) => void
+  request: (request: { method: string; params?: Array<unknown> }) => Promise<unknown>
 }
 
 export interface EIP6963ProviderDetail {
-  info: EIP6963ProviderInfo;
-  provider: EIP1193Provider;
+  info: EIP6963ProviderInfo
+  provider: EIP1193Provider
 }
 
 type EIP6963AnnounceProviderEvent = {
   detail: {
-    info: EIP6963ProviderInfo;
-    provider: EIP1193Provider;
-  };
-};
+    info: EIP6963ProviderInfo
+    provider: EIP1193Provider
+  }
+}
 
-let providers: EIP6963ProviderDetail[] = [];
+let providers: EIP6963ProviderDetail[] = []
 export const store = {
   value: () => providers,
   subscribe: (callback: () => void) => {
     function onAnnouncement(event: EIP6963AnnounceProviderEvent) {
-      if (providers.map((p) => p.info.uuid).includes(event.detail.info.uuid))
-        return;
-      providers = [...providers, event.detail];
-      callback();
+      if (providers.map(p => p.info.uuid).includes(event.detail.info.uuid)) return
+      providers = [...providers, event.detail]
+      callback()
     }
-    window.addEventListener("eip6963:announceProvider", onAnnouncement);
-    window.dispatchEvent(new Event("eip6963:requestProvider"));
+    window.addEventListener('eip6963:announceProvider', onAnnouncement)
+    window.dispatchEvent(new Event('eip6963:requestProvider'))
 
-    return () =>
-      window.removeEventListener("eip6963:announceProvider", onAnnouncement);
+    return () => window.removeEventListener('eip6963:announceProvider', onAnnouncement)
   },
-};
+}
