@@ -1,55 +1,52 @@
-import { useState, useEffect } from "react";
-import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import { GET_OPTIONS } from "../lib/constants";
+import { useState, useEffect } from 'react'
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext'
+import { GET_OPTIONS } from '../lib/constants'
 
 export default function useUser() {
-  const { siteConfig } = useDocusaurusContext();
+  const { siteConfig } = useDocusaurusContext()
   const { DASHBOARD_URL } = siteConfig?.customFields || {}
-  const [user, setUser] = useState(undefined);
-  const [loading, setLoading] = useState(true);
-  const [keys, setKeys] = useState([]);
+  const [user, setUser] = useState(undefined)
+  const [loading, setLoading] = useState(true)
+  const [keys, setKeys] = useState([])
 
   const getUserInfo = async () => {
-    setLoading(true);
+    setLoading(true)
     try {
-      const res = await fetch(`${DASHBOARD_URL}/api/me`, GET_OPTIONS);
-      const response = await res.json();
+      const res = await fetch(`${DASHBOARD_URL}/api/me`, GET_OPTIONS)
+      const response = await res.json()
       if (response?.data) {
-        setUser(response.data);
-        const userId = response.data.id;
-        const upData = await fetch(
-          `${DASHBOARD_URL}/api/v1/users/${userId}/projects`,
-          GET_OPTIONS
-        );
+        setUser(response.data)
+        const userId = response.data.id
+        const upData = await fetch(`${DASHBOARD_URL}/api/v1/users/${userId}/projects`, GET_OPTIONS)
         if (upData.ok) {
-          const upProjects = await upData.json();
-          const keysArr = upProjects?.result?.projects;
-          const allKeys = [];
+          const upProjects = await upData.json()
+          const keysArr = upProjects?.result?.projects
+          const allKeys = []
           if (keysArr) {
-            Object.keys(keysArr).forEach((key) => {
-              allKeys.push(keysArr[key]);
-            });
+            Object.keys(keysArr).forEach(key => {
+              allKeys.push(keysArr[key])
+            })
           }
-          setKeys([...allKeys]);
+          setKeys([...allKeys])
         }
       }
       if (response?.error) {
-        setUser(undefined);
-        setKeys([]);
+        setUser(undefined)
+        setKeys([])
       }
     } catch (e) {
-      setUser(undefined);
-      setKeys([]);
+      setUser(undefined)
+      setKeys([])
     } finally {
-      setLoading(false);
+      setLoading(false)
     }
-  };
+  }
   useEffect(() => {
-    getUserInfo();
-  }, []);
+    getUserInfo()
+  }, [])
   return {
     user,
     keys,
     loading,
-  };
+  }
 }
