@@ -1,143 +1,380 @@
 export const API_LINKS = {
   blockchain: [
     {
-      name: "Make requests",
-      href: "/services/how-to/make-batch-requests",
+      name: 'Make requests',
+      href: '/services/how-to/make-batch-requests',
     },
     {
-      name: "Send a transaction",
-      href: "/services/how-to/send-a-transaction",
+      name: 'Send a transaction',
+      href: '/services/how-to/send-a-transaction',
     },
     {
-      name: "Debug and optimize transactions",
-      href: "/services/how-to/trace-transactions",
+      name: 'Debug and optimize transactions',
+      href: '/services/how-to/trace-transactions',
     },
     {
-      name: "Subscribe to events",
-      href: "/services/how-to/subscribe-to-events",
+      name: 'Subscribe to events',
+      href: '/services/how-to/subscribe-to-events',
     },
     {
-      name: "Explore all methods",
-      href: "/service/index",
+      name: 'Explore all methods',
+      href: '/service/index',
     },
   ],
   storage: [
     {
-      name: "Make requests",
-      href: "/services/how-to/use-ipfs/how-to/request-rate-limits",
+      name: 'Make requests',
+      href: '/services/how-to/use-ipfs/how-to/request-rate-limits',
     },
     {
-      name: "Manage files",
-      href: "/services/how-to/use-ipfs/manage-files",
+      name: 'Manage files',
+      href: '/services/how-to/use-ipfs/manage-files',
     },
     {
-      name: "Explore all methods",
-      href: "https://docs.ipfs.tech/reference/http/api/",
+      name: 'Explore all methods',
+      href: 'https://docs.ipfs.tech/reference/http/api/',
     },
   ],
-};
+}
 
-const LOGO_PATH = "/img/logos";
+const LOGO_PATH = '/img/logos'
 
 export const NETWORK_LINKS = [
   {
-    name: "Arbitrum",
-    href: "/services/reference/arbitrum",
+    name: 'Arbitrum',
+    href: '/services/reference/arbitrum',
     logo: `${LOGO_PATH}/logo-arbitrum.svg`,
   },
   {
-    name: "Avalanche",
-    href: "/services/reference/avalanche-c-chain",
+    name: 'Avalanche',
+    href: '/services/reference/avalanche-c-chain',
     logo: `${LOGO_PATH}/logo-avalanche.svg`,
   },
   {
-    name: "Base",
-    href: "/services/reference/base",
+    name: 'Base',
+    href: '/services/reference/base',
     logo: `${LOGO_PATH}/logo-base.svg`,
   },
   {
-    name: "Blast",
-    href: "/services/reference/blast",
+    name: 'Blast',
+    href: '/services/reference/blast',
     logo: `${LOGO_PATH}/logo-blast.svg`,
   },
   {
-    name: "BNB Smart Chain",
-    href: "/services/reference/bnb-smart-chain",
+    name: 'BNB Smart Chain',
+    href: '/services/reference/bnb-smart-chain',
     logo: `${LOGO_PATH}/logo-bnb.svg`,
   },
   {
-    name: "Celo",
-    href: "/services/reference/celo",
+    name: 'Celo',
+    href: '/services/reference/celo',
     logo: `${LOGO_PATH}/logo-celo.svg`,
   },
   {
-    name: "IPFS",
-    href: "/services/reference/ipfs",
+    name: 'IPFS',
+    href: '/services/reference/ipfs',
     logo: `${LOGO_PATH}/logo-ipfs.svg`,
   },
   {
-    name: "Ethereum",
-    href: "/services/reference/ethereum",
+    name: 'Ethereum',
+    href: '/services/reference/ethereum',
     logo: `${LOGO_PATH}/logo-eth.svg`,
   },
   {
-    name: "Linea",
-    href: "/services/reference/linea",
+    name: 'Hemi',
+    href: '/services/reference/hemi',
+    logo: `${LOGO_PATH}/logo-hemi.svg`,
+  },
+  {
+    name: 'Linea',
+    href: '/services/reference/linea',
     logo: `${LOGO_PATH}/logo-linea.svg`,
   },
   {
-    name: "Mantle",
-    href: "/services/reference/mantle",
+    name: 'MegaETH',
+    href: '/services/reference/megaeth',
+    logo: `${LOGO_PATH}/logo_megaeth.svg`,
+  },
+  {
+    name: 'Mantle',
+    href: '/services/reference/mantle',
     logo: `${LOGO_PATH}/logo-mantle.svg`,
   },
   {
-    name: "opBNB",
-    href: "/api/networks/opbnb",
+    name: 'Monad',
+    href: '/services/reference/monad',
+    logo: `${LOGO_PATH}/logo-monad.svg`,
+  },
+  {
+    name: 'opBNB',
+    href: '/services/reference/opbnb',
     logo: `${LOGO_PATH}/logo-bnb.svg`,
   },
   {
-    name: "Optimism",
-    href: "/services/reference/optimism",
+    name: 'Optimism',
+    href: '/services/reference/optimism',
     logo: `${LOGO_PATH}/logo-opt.svg`,
   },
   {
-    name: "opBNB",
-    href: "/services/reference/opbnb",
-    logo: `${LOGO_PATH}/logo-bnb.svg`,
-  },
-  {
-    name: "Palm",
-    href: "/services/reference/palm",
+    name: 'Palm',
+    href: '/services/reference/palm',
     logo: `${LOGO_PATH}/logo-palm.svg`,
   },
   {
-    name: "Polygon",
-    href: "/services/reference/polygon-pos",
+    name: 'Polygon',
+    href: '/services/reference/polygon-pos',
     logo: `${LOGO_PATH}/logo-polygon.svg`,
   },
   {
-    name: "Scroll",
-    href: "/services/reference/scroll",
+    name: 'Scroll',
+    href: '/services/reference/scroll',
     logo: `${LOGO_PATH}/logo-scroll.svg`,
   },
   {
-    name: "Starknet",
-    href: "/services/reference/starknet",
+    name: 'Solana',
+    href: '/services/reference/solana',
+    logo: `${LOGO_PATH}/logo-solana.svg`,
+  },
+  {
+    name: 'Sei',
+    href: '/services/reference/sei',
+    logo: `${LOGO_PATH}/logo-sei.svg`,
+  },
+  {
+    name: 'Starknet',
+    href: '/services/reference/starknet',
     logo: `${LOGO_PATH}/logo-starknet.svg`,
   },
   {
-    name: "Swellchain",
-    href: "/services/reference/swellchain",
+    name: 'Swellchain',
+    href: '/services/reference/swellchain',
     logo: `${LOGO_PATH}/logo-swellchain.png`,
   },
   {
-    name: "Unichain",
-    href: "/services/reference/unichain",
+    name: 'Unichain',
+    href: '/services/reference/unichain',
     logo: `${LOGO_PATH}/logo-unichain.svg`,
   },
   {
-    name: "ZKsync Era",
-    href: "/services/reference/zksync",
+    name: 'ZKsync Era',
+    href: '/services/reference/zksync',
     logo: `${LOGO_PATH}/logo-zksync.svg`,
   },
-];
+]
+
+export const API_COSTS = {
+  evm: {
+    eth_accounts: 80,
+    eth_blobBaseFee: 300,
+    eth_blockNumber: 80,
+    eth_call: 80,
+    eth_chainId: 5,
+    eth_createAccessList: 80,
+    eth_estimateGas: 300,
+    eth_feeHistory: 80,
+    eth_gasPrice: 80,
+    eth_getBalance: 80,
+    eth_getBlockByHash: 80,
+    eth_getBlockByNumber: 80,
+    eth_getBlockReceipts: 1000,
+    eth_getBlockTransactionCountByHash: 150,
+    eth_getBlockTransactionCountByNumber: 150,
+    eth_getCode: 80,
+    eth_getLogs: 255,
+    eth_getProof: 150,
+    eth_getStorageAt: 80,
+    eth_getTransactionByBlockHashAndIndex: 150,
+    eth_getTransactionByBlockNumberAndIndex: 150,
+    eth_getTransactionByHash: 80,
+    eth_getTransactionCount: 80,
+    eth_getTransactionReceipt: 80,
+    eth_getUncleByBlockHashAndIndex: 150,
+    eth_getUncleByBlockNumberAndIndex: 150,
+    eth_getUncleCountByBlockHash: 150,
+    eth_getUncleCountByBlockNumber: 150,
+    eth_getWork: 80,
+    eth_hashrate: 5,
+    eth_maxPriorityFeePerGas: 80,
+    eth_mining: 5,
+    eth_protocolVersion: 5,
+    eth_sendRawTransaction: 80,
+    eth_sign: 80,
+    eth_simulateV1: 300,
+    eth_submitWork: 80,
+    eth_subscribe: 5,
+    eth_syncing: 5,
+    eth_unsubscribe: 10,
+    net_listening: 5,
+    net_peerCount: 80,
+    net_version: 5,
+    web3_clientVersion: 80,
+  },
+  evm_trace: {
+    trace_block: 300,
+    trace_call: 300,
+    trace_callMany: 300,
+    trace_filter: 300,
+    trace_transaction: 300,
+  },
+  evm_debug: {
+    //debug_getBadBlocks: 1000,
+    debug_storageRangeAt: 1000,
+    debug_traceBlock: 1000,
+    debug_traceBlockByHash: 1000,
+    debug_traceBlockByNumber: 1000,
+    debug_traceCall: 1000,
+    debug_traceTransaction: 1000,
+  },
+  evm_subscription: {
+    logs: 300,
+    newHeads: 50,
+    newPendingTransaction: 200,
+  },
+  evm_filter: {
+    eth_getFilterChanges: 140,
+    eth_getFilterLogs: 255,
+    eth_newBlockFilter: 80,
+    eth_newFilter: 80,
+    eth_uninstallFilter: 80,
+  },
+  linea: {
+    linea_estimateGas: 100,
+    linea_getTransactionExclusionStatusV1: 160,
+  },
+  polygon: {
+    bor_getAuthor: 80,
+    bor_getCurrentProposer: 80,
+    bor_getCurrentValidators: 80,
+    bor_getRootHash: 80,
+    bor_getSignersAtHash: 80,
+  },
+  solana: {
+    // Digital Asset Standard (DAS) API
+    getAsset: 160,
+    getAssetProof: 160,
+    getAssetProofBatch: 160,
+    getAssets: 160,
+    getAssetsByAuthority: 160,
+    getAssetsByCreator: 160,
+    getAssetsByGroup: 160,
+    getAssetsByOwner: 160,
+    getAssetSignatures: 160,
+    getNftEditions: 160,
+    getTokenAccounts: 160,
+    searchAssets: 160,
+    // Core Solana methods
+    getAccountInfo: 160,
+    getBalance: 160,
+    getBlock: 160,
+    getBlockCommitment: 160,
+    getBlockHeight: 160,
+    getBlockProduction: 160,
+    getBlocks: 160,
+    getBlocksWithLimit: 160,
+    getClusterNodes: 160,
+    getEpochInfo: 160,
+    getEpochSchedule: 160,
+    getFeeForMessage: 160,
+    getFirstAvailableBlock: 160,
+    getGenesisHash: 160,
+    getHealth: 160,
+    getHighestSnapshotSlot: 160,
+    getIdentity: 160,
+    getInflationGovernor: 160,
+    getInflationRate: 160,
+    getInflationReward: 160,
+    getLargestAccounts: 160,
+    getLatestBlockhash: 160,
+    getLeaderSchedule: 160,
+    getMaxRetransmitSlot: 160,
+    getMaxShredInsertSlot: 160,
+    getMinimumBalanceForRentExemption: 160,
+    getMultipleAccounts: 160,
+    getProgramAccounts: 160,
+    getRecentPerformanceSamples: 160,
+    getRecentPrioritizationFees: 160,
+    getSignaturesForAddress: 160,
+    getSignatureStatuses: 160,
+    getSlot: 160,
+    getSlotLeader: 160,
+    getSlotLeaders: 160,
+    getStakeMinimumDelegation: 160,
+    getSupply: 160,
+    getTokenAccountBalance: 160,
+    getTokenAccountsByDelegate: 160,
+    getTokenAccountsByOwner: 160,
+    getTokenLargestAccounts: 160,
+    getTokenSupply: 160,
+    getTransaction: 160,
+    getTransactionCount: 160,
+    getVersion: 160,
+    getVoteAccounts: 160,
+    isBlockhashValid: 160,
+    minimumLedgerSlot: 160,
+    requestAirdrop: 160,
+    sendTransaction: 160,
+    simulateTransaction: 160,
+  },
+  starknet: {
+    starknet_addDeclareTransaction: 700,
+    starknet_addDeployAccountTransaction: 700,
+    starknet_addInvokeTransaction: 700,
+    starknet_blockHashAndNumber: 80,
+    starknet_blockNumber: 80,
+    starknet_call: 80,
+    starknet_chainId: 5,
+    starknet_estimateFee: 240,
+    starknet_getBlockTransactionCount: 80,
+    starknet_getBlockWithTxHashes: 80,
+    starknet_getBlockWithTxs: 80,
+    starknet_getClass: 80,
+    starknet_getClassAt: 80,
+    starknet_getClassHashAt: 80,
+    starknet_getEvents: 200,
+    starknet_getNonce: 80,
+    starknet_getStateUpdate: 80,
+    starknet_getStorageAt: 80,
+    starknet_getTransactionByBlockIdAndIndex: 80,
+    starknet_getTransactionByHash: 80,
+    starknet_getTransactionReceipt: 80,
+    starknet_pendingTransactions: 160,
+    starknet_simulateTransactions: 160,
+    starknet_specVersion: 5,
+    starknet_syncing: 5,
+  },
+  zksync: {
+    zks_estimateFee: 80,
+    zks_estimateGasL1ToL2: 80,
+    zks_getAllAccountBalances: 80,
+    zks_getBlockDetails: 80,
+    zks_getBridgeContracts: 80,
+    zks_getBytecodeByHash: 80,
+    zks_getL1BatchBlockRange: 80,
+    zks_getL1BatchDetails: 80,
+    zks_getL2ToL1LogProof: 80,
+    zks_getL2ToL1MsgProof: 80,
+    zks_getMainContract: 80,
+    zks_getProof: 80,
+    zks_getRawBlockTransactions: 80,
+    zks_getTestnetPaymaster: 80,
+    zks_getTransactionDetails: 80,
+    zks_L1BatchNumber: 80,
+    zks_L1ChainId: 80,
+  },
+  gasApi: {
+    baseFeeHistory: 80,
+    baseFeePercentile: 80,
+    busyThreshold: 80,
+    suggestedGasFees: 80,
+  },
+  bundler: {
+    eth_estimateUserOperationGas: 1980,
+    eth_getUserOperationByHash: 100,
+    eth_getUserOperationReceipt: 100,
+    eth_sendUserOperation: 4950,
+    eth_supportedEntryPoints: 10,
+    pimlico_getUserOperationGasPrice: 100,
+    pimlico_getUserOperationStatus: 50,
+    pimlico_simulateAssetChanges: 4950,
+  },
+}
