@@ -11,15 +11,14 @@ import {
   AlertSuccess,
   TransactionTable,
   Hero,
-  Maintenance,
   AlertPastActivity,
+  AlertFailedPoh,
 } from '@site/src/components/Faucet'
 import { useAlert } from 'react-alert'
 import { MetamaskProviderContext } from '@site/src/theme/Root'
 import styles from './faucet.module.scss'
 import { REQUEST_PARAMS } from '@site/src/lib/constants'
 import { AlertBalanceTooLow } from '@site/src/components/Faucet/Alerts'
-import { trackInputChangeForSegment, trackPageViewForSegment } from '@site/src/lib/segmentAnalytics'
 
 const lineaMaintenanceFlag = 'linea-maintenance-mode'
 const sepoliaMaintenanceFlag = 'sepolia-maintenance-mode'
@@ -86,7 +85,9 @@ export default function Faucet() {
       const error = faucetResponse.error
 
       if (error) {
-        if (error.code === 'balance_too_low') {
+        if (error.code === 'failed_poh') {
+          alert.error(<AlertFailedPoh />)
+        } else if (error.code === 'balance_too_low') {
           alert.error(<AlertBalanceTooLow />)
         } else if (error.code === 'transaction_count_too_low') {
           alert.error(<AlertPastActivity />)
@@ -112,11 +113,6 @@ export default function Faucet() {
 
   const handleOnInputChange = value => {
     setWalletAddress(value)
-    trackInputChangeForSegment({
-      eventName: 'Wallet address',
-      userExperience: 'B',
-      timestamp: Date.now(),
-    })
   }
 
   useEffect(() => {
@@ -138,12 +134,6 @@ export default function Faucet() {
     ldClient.on(`change:${lineaMaintenanceFlag}`, handleChangeLinea)
     ldClient.on(`change:${sepoliaMaintenanceFlag}`, handleChangeSepolia)
     ldClient.on(`change:${faucetBypassDomainFlag}`, handleFaucetBypassDomain)
-
-    trackPageViewForSegment({
-      name: 'Faucet Page',
-      path: 'developer-tools/faucet',
-      userExperience: 'B',
-    })
 
     return () => {
       ldClient.off(`change:${lineaMaintenanceFlag}`, handleChangeLinea)
@@ -168,7 +158,7 @@ export default function Faucet() {
     }
   }, [metaMaskAccount, metaMaskAccountEns])
 
-  const tabItemContent = (network: 'linea' | 'sepolia') => {
+  const tabItemContent = (network: 'linea' | 'sepolia', isMaintenance: boolean) => {
     return (
       <>
         <div className={styles.topContent}>
@@ -180,6 +170,7 @@ export default function Faucet() {
             inputValue={walletAddress}
             isLoading={isLoading}
             isLimitedUserPlan={isLimitedUserPlan}
+            isMaintenance={isMaintenance}
           />
           {transactions && (
             <TransactionTable
@@ -205,12 +196,10 @@ export default function Faucet() {
       <div className={styles.tabs}>
         <Tabs className={styles.header}>
           <TabItem className={styles.content} value="sepolia" label="Ethereum Sepolia" default>
-            {isSepoliaMaintenance && <Maintenance network="sepolia" />}
-            {ldReady ? tabItemContent('sepolia') : null}
+            {ldReady ? tabItemContent('sepolia', isSepoliaMaintenance) : null}
           </TabItem>
           <TabItem className={styles.content} value="linea" label="Linea Sepolia">
-            {isLineaMaintenance && <Maintenance network="linea" />}
-            {ldReady ? tabItemContent('linea') : null}
+            {ldReady ? tabItemContent('linea', isLineaMaintenance) : null}
           </TabItem>
         </Tabs>
       </div>
