@@ -1,21 +1,21 @@
-import * as path from "path";
+import * as path from 'path'
 
 const LDPlugin = () => {
   return {
-    name: "docusaurus-plugin-launchdarkly",
+    name: 'docusaurus-plugin-launchdarkly',
     getClientModules() {
-      return [path.resolve(__dirname, "./ldClient")];
+      return [path.resolve(__dirname, './ldClient')]
     },
     configureWebpack() {
       return {
         resolve: {
           alias: {
-            launchdarkly: path.resolve(__dirname, "./ldClient.ts"),
+            launchdarkly: path.resolve(__dirname, './ldClient.ts'),
           },
         },
-      };
+      }
     },
-  };
-};
+  }
+}
 
-export default LDPlugin;
+export default LDPlugin
