@@ -1,15 +1,15 @@
-import ExecutionEnvironment from "@docusaurus/ExecutionEnvironment";
-import siteConfig from "@generated/docusaurus.config";
-import * as Sentry from "@sentry/browser";
+import ExecutionEnvironment from '@docusaurus/ExecutionEnvironment'
+import siteConfig from '@generated/docusaurus.config'
+import * as Sentry from '@sentry/browser'
 
-const isProd = process.env.NODE_ENV === "production";
+const isProd = process.env.NODE_ENV === 'production'
 
 export default (function () {
   if (!ExecutionEnvironment.canUseDOM) {
-    return null;
+    return null
   }
 
-  const { SENTRY_KEY } = siteConfig.customFields;
+  const { SENTRY_KEY } = siteConfig.customFields
 
   Sentry.init({
     dsn: SENTRY_KEY as string,
@@ -18,11 +18,11 @@ export default (function () {
     sampleRate: isProd ? 0.25 : 0,
     tracesSampleRate: 0,
     debug: !isProd,
-  });
+  })
 
   Sentry.replayIntegration({
     maskAllText: false,
-  });
+  })
 
-  return Sentry;
-})();
+  return Sentry
+})()
