@@ -1,49 +1,52 @@
-import React from "react";
-import clsx from "clsx";
-import { useWindowSize } from "@docusaurus/theme-common";
-import { useDoc } from '@docusaurus/plugin-content-docs/client';
-import DocItemPaginator from "@theme/DocItem/Paginator";
-import DocVersionBanner from "@theme/DocVersionBanner";
-import DocItemFooter from "@theme/DocItem/Footer";
-import DocItemTags from "@theme/DocItem/Tags";
-import DocItemTOCMobile from "@theme/DocItem/TOC/Mobile";
-import DocItemTOCDesktop from "@theme/DocItem/TOC/Desktop";
-import DocItemContent from "@theme/DocItem/Content";
-import DocBreadcrumbs from "@theme/DocBreadcrumbs";
-import styles from "./styles.module.css";
-import GetFeedback from "./GetFeedback.jsx";
+import React from 'react'
+import clsx from 'clsx'
+import { useWindowSize } from '@docusaurus/theme-common'
+import { useDoc } from '@docusaurus/plugin-content-docs/client'
+import DocItemPaginator from '@theme/DocItem/Paginator'
+import DocVersionBanner from '@theme/DocVersionBanner'
+import DocItemFooter from '@theme/DocItem/Footer'
+import DocItemTags from '@theme/DocItem/Tags'
+import DocItemTOCMobile from '@theme/DocItem/TOC/Mobile'
+import DocItemTOCDesktop from '@theme/DocItem/TOC/Desktop'
+import DocItemContent from '@theme/DocItem/Content'
+import DocBreadcrumbs from '@theme/DocBreadcrumbs'
+import CopyPageButton from '@site/src/components/CopyPageButton'
+import DocH1CopyPageWrapper from '@site/src/components/DocH1CopyPageWrapper'
+import styles from './styles.module.css'
 /**
  * Decide if the toc should be rendered, on mobile or desktop viewports
  */
 function useDocTOC() {
-  const { frontMatter, toc } = useDoc();
-  const windowSize = useWindowSize();
-  const hidden = frontMatter.hide_table_of_contents;
-  const canRender = !hidden && toc.length > 0;
-  const mobile = canRender ? <DocItemTOCMobile /> : undefined;
+  const { frontMatter, toc } = useDoc()
+  const windowSize = useWindowSize()
+  const hidden = frontMatter.hide_table_of_contents
+  const canRender = !hidden && toc.length > 0
+  const mobile = canRender ? <DocItemTOCMobile /> : undefined
   const desktop =
-    canRender && (windowSize === "desktop" || windowSize === "ssr") ? (
+    canRender && (windowSize === 'desktop' || windowSize === 'ssr') ? (
       <DocItemTOCDesktop />
-    ) : undefined;
+    ) : undefined
   return {
     hidden,
     mobile,
     desktop,
-  };
+  }
 }
 export default function DocItemLayout({ children }) {
-  const docTOC = useDocTOC();
+  const docTOC = useDocTOC()
   return (
     <div className="row">
-      <div className={clsx("col", !docTOC.hidden && styles.docItemCol)}>
+      <div className={clsx('col', !docTOC.hidden && styles.docItemCol)}>
         <DocVersionBanner />
         <div className={styles.docItemContainer}>
           <article>
             <DocBreadcrumbs />
             <DocItemTags />
             {docTOC.mobile}
+            <DocH1CopyPageWrapper />
+            <CopyPageButton />
             <DocItemContent>{children}</DocItemContent>
-            {/* <GetFeedback /> */}
+
             <DocItemFooter />
           </article>
           <DocItemPaginator />
@@ -51,5 +54,5 @@ export default function DocItemLayout({ children }) {
       </div>
       {docTOC.desktop && <div className="col col--3">{docTOC.desktop}</div>}
     </div>
-  );
+  )
 }
