@@ -1,3 +1,3 @@
 export default function PaginatorWrapper() {
-  return null;
+  return null
 }
