@@ -1,16 +1,16 @@
-import React from "react";
-import DocSidebarItem from "@theme-original/DocSidebarItem";
-import type DocSidebarItemType from "@theme/DocSidebarItem";
-import type { WrapperProps } from "@docusaurus/types";
-import clsx from "clsx";
+import React from 'react'
+import DocSidebarItem from '@theme-original/DocSidebarItem'
+import type DocSidebarItemType from '@theme/DocSidebarItem'
+import type { WrapperProps } from '@docusaurus/types'
+import clsx from 'clsx'
 
-import styles from "./index.module.css";
+import styles from './index.module.css'
 
-type Props = WrapperProps<typeof DocSidebarItemType>;
+type Props = WrapperProps<typeof DocSidebarItemType>
 
 export default function DocSidebarItemWrapper(props: Props): JSX.Element {
-  if (props.item.type === "category" || props.item.type === "link") {
-    const { className, ...itemProps } = props.item;
+  if (props.item.type === 'category' || props.item.type === 'link') {
+    const { className, ...itemProps } = props.item
 
     return (
       <>
@@ -24,11 +24,11 @@ export default function DocSidebarItemWrapper(props: Props): JSX.Element {
           }}
         />
       </>
-    );
+    )
   }
   return (
     <>
       <DocSidebarItem {...props} />
     </>
-  );
+  )
 }
