@@ -1,9 +1,9 @@
 // Import the original mapper
-import Pill from "@site/src/components/Pill";
-import MDXComponents from "@theme-original/MDXComponents";
+import Pill from '@site/src/components/Pill'
+import MDXComponents from '@theme-original/MDXComponents'
 
 export default {
   // Re-use the default mapping
   ...MDXComponents,
   Pill,
-};
+}
