@@ -1,7 +1,7 @@
-import ComponentTypes from '@theme-original/NavbarItem/ComponentTypes';
-import NavbarWallet from '@site/src/components/NavbarWallet';
+import ComponentTypes from '@theme-original/NavbarItem/ComponentTypes'
+import NavbarWallet from '@site/src/components/NavbarWallet'
 
 export default {
-    ...ComponentTypes,
-    'custom-navbarWallet': NavbarWallet,
-};
+  ...ComponentTypes,
+  'custom-navbarWallet': NavbarWallet,
+}
