@@ -1,18 +1,9 @@
-import React, {
-  ReactElement,
-  createContext,
-  useEffect,
-  useState,
-  useCallback,
-} from "react";
-import { Provider as AlertProvider } from "react-alert";
-import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import { AlertTemplate, options } from "@site/src/components/Alert";
-import { MetaMaskSDK, SDKProvider } from "@metamask/sdk";
-import {
-  REF_ALLOW_LOGIN_PATH,
-  REQUEST_PARAMS,
-} from "@site/src/lib/constants";
+import React, { ReactElement, createContext, useEffect, useState, useCallback } from 'react'
+import { Provider as AlertProvider } from 'react-alert'
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext'
+import { AlertTemplate, options } from '@site/src/components/Alert'
+import { MetaMaskSDK, SDKProvider } from '@metamask/sdk'
+import { REF_ALLOW_LOGIN_PATH, REQUEST_PARAMS } from '@site/src/lib/constants'
 import {
   clearStorage,
   getUserIdFromJwtToken,
@@ -21,52 +12,52 @@ import {
   getUksTier,
   AUTH_WALLET_PROJECTS,
   AUTH_WALLET_ENS,
-  getWalletEns
-} from "@site/src/lib/siwsrp/auth";
+  getWalletEns,
+} from '@site/src/lib/siwsrp/auth'
 import AuthModal, {
   AUTH_LOGIN_STEP,
   WALLET_LINK_TYPE,
-} from "@site/src/components/AuthLogin/AuthModal";
+} from '@site/src/components/AuthLogin/AuthModal'
 
 interface Project {
-  id: string;
-  userId: string;
-  name: string;
-  created: number;
-  updated: number;
-  deleted: boolean;
-  settings: any;
+  id: string
+  userId: string
+  name: string
+  created: number
+  updated: number
+  deleted: boolean
+  settings: any
   networks: {
-    [key: string]: { subnets: number[] };
-  };
-  role: string;
+    [key: string]: { subnets: number[] }
+  }
+  role: string
 }
 
 interface IMetamaskProviderContext {
-  token: string;
-  projects: { [key: string]: Project };
-  setProjects: (arg: { [key: string]: Project }) => void;
-  metaMaskDisconnect: () => Promise<void>;
-  metaMaskWalletIdConnectHandler: () => Promise<void>;
-  userId: string | undefined;
-  metaMaskAccount: string;
-  metaMaskAccountEns: string;
-  setMetaMaskAccount: (arg: string[] | string) => void;
-  metaMaskProvider: SDKProvider;
-  setMetaMaskProvider: (arg: SDKProvider) => void;
-  uksTier: string;
-  sdk: MetaMaskSDK;
-  setNeedsMfa: (arg: boolean) => void;
-  needsMfa: boolean;
-  setWalletLinked: (arg: WALLET_LINK_TYPE) => void;
-  walletLinked: WALLET_LINK_TYPE | undefined;
-  setWalletAuthUrl: (arg: string) => void;
-  walletAuthUrl: string;
-  userAPIKey?: string;
-  setUserAPIKey?: (key: string) => void;
-  fetchLineaEns?: (walletId: string) => Promise<void>;
-  userEncPublicKey?: string;
-  setUserEncPublicKey?: (key: string) => void;
+  token: string
+  projects: { [key: string]: Project }
+  setProjects: (arg: { [key: string]: Project }) => void
+  metaMaskDisconnect: () => Promise<void>
+  metaMaskWalletIdConnectHandler: () => Promise<void>
+  userId: string | undefined
+  metaMaskAccount: string
+  metaMaskAccountEns: string
+  setMetaMaskAccount: (arg: string[] | string) => void
+  metaMaskProvider: SDKProvider
+  setMetaMaskProvider: (arg: SDKProvider) => void
+  uksTier: string
+  sdk: MetaMaskSDK
+  setNeedsMfa: (arg: boolean) => void
+  needsMfa: boolean
+  setWalletLinked: (arg: WALLET_LINK_TYPE) => void
+  walletLinked: WALLET_LINK_TYPE | undefined
+  setWalletAuthUrl: (arg: string) => void
+  walletAuthUrl: string
+  userAPIKey?: string
+  setUserAPIKey?: (key: string) => void
+  fetchLineaEns?: (walletId: string) => Promise<void>
+  userEncPublicKey?: string
+  setUserEncPublicKey?: (key: string) => void
 }
 
 export const MetamaskProviderContext = createContext<IMetamaskProviderContext>({
@@ -88,18 +79,18 @@ export const MetamaskProviderContext = createContext<IMetamaskProviderContext>({
   setWalletLinked: () => {},
   walletLinked: undefined,
   setWalletAuthUrl: () => {},
-  walletAuthUrl: "",
-  userAPIKey: "",
+  walletAuthUrl: '',
+  userAPIKey: '',
   setUserAPIKey: () => {},
   fetchLineaEns: () => new Promise(() => {}),
   userEncPublicKey: undefined,
   setUserEncPublicKey: () => {},
-});
+})
 
 const sdk = new MetaMaskSDK({
   dappMetadata: {
-    name: "Reference pages",
-    url: "https://docs.metamask.io/",
+    name: 'Reference pages',
+    url: 'https://docs.metamask.io/',
   },
   preferDesktop: true,
   extensionOnly: true,
@@ -107,166 +98,147 @@ const sdk = new MetaMaskSDK({
   logging: {
     sdk: false,
   },
-});
+})
 
 export const LoginProvider = ({ children }) => {
-  const [projects, setProjects] = useState({});
-  const [userId, setUserId] = useState<string>("");
-  const [token, setToken] = useState<string>(undefined);
-  const [openAuthModal, setOpenAuthModal] = useState<boolean>(false);
-  const [metaMaskProvider, setMetaMaskProvider] = useState(undefined);
-  const [metaMaskAccount, setMetaMaskAccount] = useState(undefined);
-  const [metaMaskAccountEns, setMetaMaskAccountEns] = useState(undefined);
-  const [uksTier, setUksTier] = useState(undefined);
-  const [isInitialized, setIsInitialized] = useState<boolean>(false);
-  const [step, setStep] = useState<AUTH_LOGIN_STEP>(AUTH_LOGIN_STEP.CONNECTING);
-  const [walletLinked, setWalletLinked] = useState<
-    WALLET_LINK_TYPE | undefined
-  >(undefined);
-  const [needsMfa, setNeedsMfa] = useState<boolean>(false);
-  const [walletAuthUrl, setWalletAuthUrl] = useState<string>("");
-  const [userAPIKey, setUserAPIKey] = useState("");
-  const [userEncPublicKey, setUserEncPublicKey] = useState(undefined);
-  const { siteConfig } = useDocusaurusContext();
-  const { DASHBOARD_URL, GF_SURVEY_KEY, LINEA_ENS_URL } = siteConfig?.customFields || {};
+  const [projects, setProjects] = useState({})
+  const [userId, setUserId] = useState<string>('')
+  const [token, setToken] = useState<string>(undefined)
+  const [openAuthModal, setOpenAuthModal] = useState<boolean>(false)
+  const [metaMaskProvider, setMetaMaskProvider] = useState(undefined)
+  const [metaMaskAccount, setMetaMaskAccount] = useState(undefined)
+  const [metaMaskAccountEns, setMetaMaskAccountEns] = useState(undefined)
+  const [uksTier, setUksTier] = useState(undefined)
+  const [isInitialized, setIsInitialized] = useState<boolean>(false)
+  const [step, setStep] = useState<AUTH_LOGIN_STEP>(AUTH_LOGIN_STEP.CONNECTING)
+  const [walletLinked, setWalletLinked] = useState<WALLET_LINK_TYPE | undefined>(undefined)
+  const [needsMfa, setNeedsMfa] = useState<boolean>(false)
+  const [walletAuthUrl, setWalletAuthUrl] = useState<string>('')
+  const [userAPIKey, setUserAPIKey] = useState('')
+  const [userEncPublicKey, setUserEncPublicKey] = useState(undefined)
+  const { siteConfig } = useDocusaurusContext()
+  const { DASHBOARD_URL, LINEA_ENS_URL } = siteConfig?.customFields || {}
 
   if (sdk.isInitialized() && !isInitialized) {
-    setIsInitialized(true);
+    setIsInitialized(true)
   }
 
   const fetchLineaEns = async (rawAddress: string) => {
     if (getWalletEns()) {
-      setMetaMaskAccountEns(getWalletEns());
+      setMetaMaskAccountEns(getWalletEns())
     } else if (rawAddress) {
-      const address = String(rawAddress).toLowerCase();
+      const address = String(rawAddress).toLowerCase()
       try {
         const res = await (
           await fetch(LINEA_ENS_URL as string, {
-            ...REQUEST_PARAMS("POST"),
+            ...REQUEST_PARAMS('POST'),
             body: JSON.stringify({
               query: `query getNamesForAddress {domains(first: 1, where: {and: [{or: [{owner: \"${address}\"}, {registrant: \"${address}\"}, {wrappedOwner: \"${address}\"}]}, {parent_not: \"0x91d1777781884d03a6757a803996e38de2a42967fb37eeaca72729271025a9e2\"}, {or: [{expiryDate_gt: \"1721033912\"}, {expiryDate: null}]}, {or: [{owner_not: \"0x0000000000000000000000000000000000000000\"}, {resolver_not: null}, {and: [{registrant_not: \"0x0000000000000000000000000000000000000000\"}, {registrant_not: null}]}]}]}) {...DomainDetailsWithoutParent}} fragment DomainDetailsWithoutParent on Domain {name}`,
             }),
           })
-        ).json();
-        const walletEns = res.data.domains[0].name;
-        setMetaMaskAccountEns(walletEns);
-        sessionStorage.setItem(AUTH_WALLET_ENS, walletEns);
+        ).json()
+        const walletEns = res.data.domains[0].name
+        setMetaMaskAccountEns(walletEns)
+        sessionStorage.setItem(AUTH_WALLET_ENS, walletEns)
       } catch (e) {
-        setMetaMaskAccountEns(undefined);
+        setMetaMaskAccountEns(undefined)
       }
     }
-  };
+  }
 
   const getStaleDate = async () => {
     try {
-      setProjects(
-        JSON.parse(sessionStorage.getItem(AUTH_WALLET_PROJECTS) || "{}"),
-      );
-      setUserId(getUserIdFromJwtToken());
-      setToken(getTokenString());
-      setUksTier(getUksTier());
-      setMetaMaskAccountEns(getWalletEns());
-      const accounts = await sdk.connect();
-      setMetaMaskAccount(accounts);
+      setProjects(JSON.parse(sessionStorage.getItem(AUTH_WALLET_PROJECTS) || '{}'))
+      setUserId(getUserIdFromJwtToken())
+      setToken(getTokenString())
+      setUksTier(getUksTier())
+      setMetaMaskAccountEns(getWalletEns())
+      const accounts = await sdk.connect()
+      setMetaMaskAccount(accounts)
       if (accounts && accounts.length > 0) {
-        setMetaMaskAccount(accounts[0]);
-        const provider = sdk.getProvider();
-        setMetaMaskProvider(provider);
+        setMetaMaskAccount(accounts[0])
+        const provider = sdk.getProvider()
+        setMetaMaskProvider(provider)
       }
     } catch (e) {}
-  };
+  }
 
   const metaMaskWalletIdConnectHandler = useCallback(async () => {
     try {
-      setOpenAuthModal(true);
+      setOpenAuthModal(true)
     } catch (err) {
-      console.warn("failed to connect..", err);
+      console.warn('failed to connect..', err)
     }
-  }, [setOpenAuthModal]);
+  }, [setOpenAuthModal])
 
   const metaMaskDisconnect = useCallback(async () => {
     try {
-      await sdk?.terminate();
-      setOpenAuthModal(false);
-      setUserId(undefined);
-      setToken(undefined);
-      setMetaMaskAccount(undefined);
-      setMetaMaskAccountEns(undefined);
-      setUksTier(undefined);
-      setProjects({});
-      setWalletLinked(undefined);
-      setNeedsMfa(false);
-      setUserAPIKey("");
-      clearStorage();
+      await sdk?.terminate()
+      setOpenAuthModal(false)
+      setUserId(undefined)
+      setToken(undefined)
+      setMetaMaskAccount(undefined)
+      setMetaMaskAccountEns(undefined)
+      setUksTier(undefined)
+      setProjects({})
+      setWalletLinked(undefined)
+      setNeedsMfa(false)
+      setUserAPIKey('')
+      clearStorage()
     } catch (err) {
-      console.warn("failed to disconnect..", err);
+      console.warn('failed to disconnect..', err)
     }
-  }, [
-    sdk,
-    setOpenAuthModal,
-    setUserId,
-    setToken,
-    setMetaMaskAccount,
-    setUksTier,
-    setProjects,
-  ]);
+  }, [sdk, setOpenAuthModal, setUserId, setToken, setMetaMaskAccount, setUksTier, setProjects])
 
   useEffect(() => {
-    const provider = sdk?.getProvider();
-    setMetaMaskProvider(provider);
-    if ((window as any)?.usabilla && window?.innerWidth > 1720) {
-      (window as any)?.usabilla?.load("w.usabilla.com", GF_SURVEY_KEY);
-    }
-  }, []);
+    const provider = sdk?.getProvider()
+    setMetaMaskProvider(provider)
+  }, [])
 
   useEffect(() => {
     if (isInitialized && sdk.isExtensionActive()) {
-      const provider = sdk.getProvider();
-      sdk.resume();
-      setMetaMaskProvider(provider);
+      const provider = sdk.getProvider()
+      sdk.resume()
+      setMetaMaskProvider(provider)
     }
-  }, [isInitialized]);
+  }, [isInitialized])
 
   useEffect(() => {
-    const url = new URL(window.location.href);
-    getStaleDate();
-    if (REF_ALLOW_LOGIN_PATH.some((item) => url.pathname.includes(item))) {
-      const token = url.searchParams.get("token");
+    const url = new URL(window.location.href)
+    getStaleDate()
+    if (REF_ALLOW_LOGIN_PATH.some(item => url.pathname.includes(item))) {
+      const token = url.searchParams.get('token')
 
       if (token) {
-        saveTokenString(token);
-        setToken(token);
-        const userIdFromjwtToken = getUserIdFromJwtToken();
-        setUserId(userIdFromjwtToken);
-
-        (async () => {
+        saveTokenString(token)
+        setToken(token)
+        const userIdFromjwtToken = getUserIdFromJwtToken()
+        setUserId(userIdFromjwtToken)
+        ;(async () => {
           try {
             const projectsResponse = await fetch(
               `${DASHBOARD_URL}/api/v1/users/${userIdFromjwtToken}/projects`,
               {
-                ...REQUEST_PARAMS("GET", { Authorization: `Bearer ${token}` }),
-              },
-            );
-            const res = await projectsResponse.json();
-            if (res.error) throw new Error(res.error.message);
+                ...REQUEST_PARAMS('GET', { Authorization: `Bearer ${token}` }),
+              }
+            )
+            const res = await projectsResponse.json()
+            if (res.error) throw new Error(res.error.message)
 
             const {
               result: { projects },
-            } = res;
-            sessionStorage.setItem(
-              AUTH_WALLET_PROJECTS,
-              JSON.stringify(projects),
-            );
-            setProjects(projects);
-            window.location.replace(`${url.origin}${url.pathname}`);
+            } = res
+            sessionStorage.setItem(AUTH_WALLET_PROJECTS, JSON.stringify(projects))
+            setProjects(projects)
+            window.location.replace(`${url.origin}${url.pathname}`)
           } catch (e: any) {
-            setStep(AUTH_LOGIN_STEP.CONNECTION_ERROR);
-            setOpenAuthModal(true);
+            setStep(AUTH_LOGIN_STEP.CONNECTION_ERROR)
+            setOpenAuthModal(true)
           }
-        })();
+        })()
       }
     }
-  }, []);
+  }, [])
 
   return (
     <MetamaskProviderContext.Provider
@@ -297,8 +269,7 @@ export const LoginProvider = ({ children }) => {
           userEncPublicKey,
           setUserEncPublicKey,
         } as IMetamaskProviderContext
-      }
-    >
+      }>
       {children}
       <AuthModal
         open={openAuthModal}
@@ -310,8 +281,8 @@ export const LoginProvider = ({ children }) => {
         step={step}
       />
     </MetamaskProviderContext.Provider>
-  );
-};
+  )
+}
 
 export default function Root({ children }: { children: ReactElement }) {
   return (
@@ -320,5 +291,5 @@ export default function Root({ children }: { children: ReactElement }) {
         {children}
       </AlertProvider>
     </LoginProvider>
-  );
+  )
 }
